[IMP] l10n_mx/saas~16.2.2.0: adjusting DIOT tag migration
In odoo/odoo#204742 we recently added new tax tags for the DIOT report in 16.0
(needed to comply with a legal change).
Since the DIOT report had been refactored in 17.0 to use the tax_tags engine
(instead of a custom engine), there was already an existing upgrade script to
update the tax tags on account.move.line records.
This PR modifies the upgrade script so that it can correctly update the new tax
tags on AMLs.
However, until the 17.0 forward-port is merged, the new tax tags don't exist
yet in 17.0, so we need to make the upgrade fail if any of the new tags are
used.
See odoo/odoo#213194
See odoo/enterprise#87102

closes odoo/upgrade#7841

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>