[FIX] account_reports/saas~18.3: Execute Cron to generate returns
When this https://github.com/odoo/enterprise/pull/85509 was merged,
we added a nextcall to the cron that generated the returns.

This can be a problem as users migrating have to wait the date of the
next call to be able to get their returns created.

We should run the cron during migration manually to avoid this.

closes odoo/upgrade#7869

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>