[IMP] account/saas~18.3: improve fp tax mappings
After testing the production migration, we found improvements for migrating
databases with questionable data.

Before:
1. All source taxes in account_fiscal_position_tax were considered to be
domestic, but when those taxes are also destination taxes (due to
misconfiguration) they should probably not be in the domestic fiscal position.
Except...
2. Destination taxes that map to itself should also be included in the list of
domestic taxes. (This is used by l10n_es to add a withholding tax to a
domestic tax) see image on PR.
3. Fiscal positions that had no mapping (for eg. EU B2C) were not added to the
list of fiscal positions a tax belongs to. Using this fiscal position removes
taxes.

After:
- Limit/ reduce domestic taxes
- Link domestic taxes to Empty Fiscal Positions

task-4863170

closes odoo/upgrade#7872

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>