[FIX] base/saas~18.4: force upgrade module
The module l10n_ch_hr_payroll_account is not a direct dependency of
l10n_ch_hr_payroll_elm_transmission_account, it is there only by
auto-install. But it will be force installed when we merge the modules
and the first one is installed. This will not run the pre-script that
removes 2 views from the old module.

Part-of: odoo/upgrade#7878
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>