[FIX] hr_contract/tests: set null contract_type_id
In the test prepare when creating contracts, some contracts
so not specify the contract type assuming it will be None
and the test_check validates it. In the case of l10n modules
installed, the contract type may have a default value which
makes the check fail. The type must be explicitly set to None.

Part-of: odoo/upgrade#7878
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>