[FIX] account_disallowed_expenses{_fleet}/saas~18.{2,4}: Rate migrations
removal

We considered that the added deductibility of 20% for invoicing software
expenses was a 120% Disallowed Expense, but this is wrong and that does
not exist.
It is just a 20% rate on particular software expense that can be
additionally reported in a specific section of the Fiscal Declaration,
and that will further reduce the income before the computation of the
income tax (the other 100% part being reported normally).

odoo/enterprise#88049
task-4848518

closes odoo/upgrade#7879

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>