[ADD] base/0.0.0: forbid the ORM to alter the table of some models
On some models, using the ORM to compute the value of the new
computed-stored fields can be long and lead to MemoryError.
Such fields needs to be computed in SQL or via `util.recompute_fields`.

This also the case for new columns with a default values.

When a column change type, the ORM do a plain `ALTER TYPE` query, which
can be slow. `util.alter_column_type` should be used as it has several
optimizations,

Logs a critical message to fail the upgrade on CI.