[IMP] appointment/saas~18.5: m2m questions, new 'phone' type and options
A few model changes are made regarding questions in appointment:

1. Make questions a m2m field on appointment.type

In order to allow users to use questions across similar appointments,
we now make them a 'library' of questions, by changing question_ids to
a m2m field instead of a o2m.

On the appointment.question model, appointment_type_id (m2o) is now
appointment_type_ids (m2m). The relational table between appointments
and questions is created and populated.

2. Add is_reusable boolean field on appointment.question

In order to let the end user configure the fact that a question can be
added in all appointments or is exclusive, we add is_reusable. While
True by default in the model, we want to make it False on existing records
in the migration, so that they do not appear as candidates when adding a
question in the appointment.type form. They will start with a 'clean
sheet', so to speak.

3. Add is_default boolean field on appointment.question

Allow the user to determine whether a question should be added by default
when creating a new appointment_type. is_default is False by default.

4. Add a new 'phone' question_type and a phone question data record

Before, the phone was hard-coded in the attendee form in the front-end,
and was required. Now, it is removed, and is instead implemented with
an appointment.question. It has a new question_type 'phone', and could
be not used, in order to allow phoneless appointment.type, which was
not possible before.

We add a phone question to make sure all appointment.type still have
the phone question in the front-end form after the migration. We also
link it to a data, that will be installed with the module, or created
on migration.

-> We link all existing appointment types to that question and make sure
that it comes first in sequence, so that this will be at the top of the
questions, as it was before migration in the front-end form.

The question is both reusable and a default question, unlike existing
records before migration.

5. Make question_type required on appointment.question

Instead of removing the records with no question_type, we instead set a
default 'char' or 'select' when it has some suggested answers. (see
answer_ids field)

A test is included.

ENT PR: https://github.com/odoo/enterprise/pull/82973

Task-4655040

closes odoo/upgrade#7691

Related: odoo/enterprise#82973
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>