[FIX] account_accountant/tests: make test independent
The test validates that a partner on a statement can be matched
from similar previous statements using similarities on account
numbers. The test also uses the main company which leads to
account numbers matching other demo data and failing the test
based on specific modules installed.
This fix is making the test use a new company to avoid
interferences with other tests and demo.

closes odoo/upgrade#7878

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>