[FIX] l10n_mx/16.0: match unique groups
When we load new account groups during the upgrade there may be some
collisions with already existing groups on the client's DB. The
constraint is implemented in Python side
https://github.com/odoo/odoo/blob/9728a1754424b9e5bfa93ec5d53a5179fa9d7d05/addons/account/models/account_account.py#L825

In this patch we propose to reuse existing groups with same prefix and
suffix per company. Technically, we prefer to leverage the `_match_uniq`
battle tested mechanism. For this we add an index with the constraints
and remove it as soon as the load of the new groups are done.

There is a chance that the new index cannot be added in some DBs. Since
the Python constraint is stricter than the unique index, records
violating the index are already wrong thus extra upgrade failures would
be correct.

Example existing data in the DB:
```
=> select company_id,name,code_prefix_start,code_prefix_end from account_group where name in ('PASIVO','CAJA')
+------------+--------+-------------------+-----------------+
| company_id | name   | code_prefix_start | code_prefix_end |
|------------+--------+-------------------+-----------------|
| 1          | PASIVO | 2                 | 2               |
| 1          | CAJA   | 101               | 101             |
+------------+--------+-------------------+-----------------+
```

Values in standard (16.0, same order):
https://github.com/odoo/odoo/blob/80a12130bed0d18da99e41d48b3f0ca937ffddf9/addons/l10n_mx/data/account.group.template.csv?plain=1#L251
https://github.com/odoo/odoo/blob/80a12130bed0d18da99e41d48b3f0ca937ffddf9/addons/l10n_mx/data/account.group.template.csv?plain=1#L4

Note that the name is not 100% the same (singular vs plural, and capital
vs lowercase) but still with the same meaning.

upg-2986707

closes odoo/upgrade#7947

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>