[FIX] mail,{*}/saas~18.5: reset template lang
The mail lang is now dynamically chosen to be
the customer's lang but the mail template records
are not updated. Some of them are noupdate and those
that get updated do not reset the langauge when updated.
Here we update the langauge for updated records
if it is a standard value.
The description of a noupdate calendar template was
also modified.

https://github.com/odoo/odoo/pull/189743
https://github.com/odoo/enterprise/pull/81862

closes odoo/upgrade#7954

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>