[IMP] documents/saas~18.5: improve share panel
In odoo/enterprise#87971, we extend the wizard allowing to share documents with
partners:
- to works on multiple documents (instead of one)
- to also support the edition of access rights (on multiple documents) as a
replacement/improvement of the owl implementation

To keep git history, we have renamed the documents.access.invite wizard into
documents.sharing and based the new implementation on it. Here we just remove
the renamed model as the related view will be automatically deleted. We also
force the update of the notification email sent when sharing files with someone
as it now deals with multiple documents instead of a single one.

Task-4848234

Part-of: odoo/upgrade#7959
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>