[IMP] base/tests: handle multi-company during mock views
After this commit: odoo/odoo@6213c40932236101b529b82f0ea9fce1829c8c24,
If a record's `company_id` is accessible by the user
(i.e., present in `self.env.user.company_ids`), but not in the current
`allowed_company_ids`, the frontend silently adds it to the context
and successfully opens the form view.

However, during upgrade mocking in test scenarios, the lack of a UI means that
the `AccessError` is not gracefully handled, and the upgrade fails.

This commit mirrors the frontend logic in `v18`: when a record read raises
an `AccessError`, we check if a `suggested_company` is present in the
exception's context (as provided by `_get_redirect_suggested_company`).
If the company is among the user's allowed companies but missing from the
current context, we update `allowed_company_ids` and retry the read.

opw-4886199
upg-2989592

Part-of: odoo/upgrade#7960
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>