[FIX] point_of_sale/12.0: remove demo data
In the refactor of product demo data done in 12.0 the barcodes where
kept for new products. This then causes an issue when adding the new
demo products since the barcode collides. To avoid the error we remove
the old demo products.

This patch is mostly to please the CI since it has low impact on
client's DBs.

See odoo/odoo@bed29a4aeab0ec6ef0c02975ee86655d6a030299

Part-of: odoo/upgrade#7970
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>