[FIX] hr/saas~18.4: update record rule
The record rule ir_rule_hr_contract_manager is in no_update in the data,
but the group has changed and old group is deleted. This leads to get
the rule as global, because the old group deletion set it to null.

Introduced in PR: https://github.com/odoo/odoo/pull/202869

closes odoo/upgrade#7986

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>