[FIX] account_reports/saas~18.3: generate account returns
upg-2998822

In 1487b6434e7e242d430fabbb57b60395f4b8aba0 we introduced a cron trigger to
ensure an early execution of the cronjob wrt the end of the upgrade as account
returns need to be populated for the module to work properly and the cron could
take days to be triggered by default.

However, cronjobs are disabled for test upgrades
https://github.com/odoo/upgrade-platform/blob/2767f6d0bd4fbcf85de8666dd68622853815b064/upgrade-docker/sanitizer.py#L487,
therefore even the cron trigger does not fire, resulting in broken features, at
upgrade db testing time. Instead we simply run the code to generate account
returns.

closes odoo/upgrade#7982

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>