[FIX] documents/saas~17.5: force alias_force_thread_id to null
Having this value set to the document folder_id results
in new incoming emails sent to this alias
creating mail.message records.
This action populates the chatter with the message body
and adds attachments to the attached files,
instead of generating new documents.

By forcing the value to null, we ensure new incoming emails
create documents.document, as expected.

opw-4853118

closes odoo/upgrade#8010

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>