[IMP] appointment/saas~18.5: redefine a few fields
In order to split different configuration logics, we move
from a single assign_method selection field to two boolean fields
is_auto_assign (manual selection if False) and is_date_first (it
will be resource selection first if False).

Also move from a default-false appointment_manual_confirmation to
a default-true auto_confirm option. The logic is now opposite: when
the option is enabled, then we auto-confirm based on the percentage.
If 100% is set, we consider we can always confirm the booking. If
auto-confirm is false, meeting status is always 'request'.

-> For upgrade, we keep the same percentage when manual_confirmation
and manage_capacity were true, as 'manual_confirmation above x %'
is now 'auto_confirm up to x%'. The rest of records have their
manual confirmation percentage set to the new default 100.0.

-> The only case we set auto_confirm to False in the migration is
when the manual_confirmation was True, but manage_capacity was False,
as in that case the percentage had no impact and the meetings were
always in 'request' status. Note that now, the percentage will be used
as well regardless of the manage_capacity value.

The selection field avatars_display is also changed into a boolean
field show_avatars.

A simple domain adapter is included for assign_method.

ENT PR: odoo/enterprise#88637

Task-4898893

closes odoo/upgrade#8014

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>