[FIX] hr/saas~18.4: merge contract groups into hr groups
Before this commit, the contract groups and all their references are not
replaced by the hr groups as we did in community and enterprise during
the refactoring made in 18.4.

This commit makes sure the contract groups are correctly merged into hr
groups and their references are replaced by the hr groups ones.
This commit also removes the records removed in the security folder of
contract since those records are in no update and potentially not
removed during the migration.

Part-of: odoo/upgrade#8015
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>