[IMP] l10n_my_edi*/saas~18.5: cleaning up
When recently implementing support for the PoS,
we started to plan an improvement of the whole
feature by implementing a MyInvois Document model
to regroup the flows in one place.

In this improvement, we are finalizing what we
started by implementing this new module with the
invoicing app as well so that we can reduce the
amount of duplicated code both in all the 'sending'
flows but also in the xml generation code.

task-4929885
see https://github.com/odoo/odoo/pull/220034

closes odoo/upgrade#8107

Related: odoo/odoo#220034
Signed-off-by: William André (wan) <wan@odoo.com>