[FIX] l10n_it_edi/0.0.0: fix duplicate code in l10n_it.document.type
Prior to this commit odoo/odoo@82e014702d472ee164f69623055c96cce80c5f6f,
it was possible to create duplicate document types.
This commit introduces a constraint to prevent such duplication.
As this fix targets a stable version, an upgrade script is included to
automatically update existing duplicate records and make their codes unique.
The main purpose is to notify customers of their faulty data so they can review
and correct it as needed.

Steps to Reproduce:
- Install the `l10n_it_edi_ndd` module in version 18.0
- create a new invoice and confirm it.
- from `Electronic Invoicing` page, Open Document Type and duplicate it.

observation: the database gets blocked when migrating to saas-18.2 .

opw - 4902513
upg - 2995474

closes odoo/upgrade#8032

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>