[IMP] {l10n_be,account}_reports/saas~18.5: tax returns improvements
In the related enterprise PR, we improve the views and model of
the tax returns. for the return checks, we remove the chatter and
some views.

VAT returns now also have 4 stages instead of 3 before. The "Submit"
stage is split into 2, making necessary an upgrade script to handle
the 2 wizards that now exist.

The field approver_ids on the return checks is also removed. It could
contain several users. This field is replaced by 2 new fields, approver_id
and supervisor_id. For the migration, if an admin was present in the
original field, we set it as both the approver and supervisor, and
if only a bookkeeper was present, it is set as approver.

enterprise-https://github.com/odoo/enterprise/pull/89384
task-4891073

closes odoo/upgrade#8035

Related: odoo/enterprise#89384
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>