[FIX] im_livechat/saas~18.4: agent report compute stored fields
This commit computes values for several computed stored fields that
were added to the live chat module for reporting:
- `livechat_agent_{providing/requesting}_help_history` (discuss_channel):
If more than one agent joined the chat, we consider the first one to have
requested help and the last one to have provided help.
- `livechat_outcome` (discuss_channel): Single value for
`livechat_is_escalated` and `livechat_failure`, used in group by.
- `livechat_start_hour` (discuss_channel): start hour of the session,
computed from `create_date` and used in group by.
- `livechat_week_day` (discuss_channel): computed from `create_date` and
used in group by.
- `help_status` (history): Selection based on the discuss channel fields
(requesting/providing help), used in group by.

Other fields are not computed fields but still need to be filled in
order to be used in reporting views:
- `session_duration_hour` (history): Duration of the session for an agent. The
session starts at the `create_date` of the agent history (when the
agent joined the channel). If the chat is ongoing, the end time is the
date of the last message. If the chat has ended, the end time is
`livechat_end_dt`.
- `message_count`: Number of messages sent from a given visitor/agent on
a live chat.

community: https://github.com/odoo/odoo/pull/213340,
           https://github.com/odoo/odoo/pull/214795.

closes odoo/upgrade#7922

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>