[FIX] base/0.0.0: don't remove res_id=0 attachments
When an html field is used with the class `oe-bordered-editor` in a form
view for new records users can create attachments directly. The
attachments are inserted with res_id=0 and used inline in the html value
of the field. Since we cannot know if a `res_id=0` attachment was used
from an html field we cannot really remove them. Thus when cleaning
transients we need to keep the attachments.

opw-4812659

closes odoo/upgrade#8039

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>