[IMP] l10n_mx/saas~16.2.2.0: adjusting DIOT tax tag migration
In odoo/odoo#218230 we recently added new tax tags for the DIOT report in 16.0.
Even though this change was implemented a few days ago,
the payment columns should be based on the base amount and not the tax amount,
so even more tags are added.
This PR modifies the upgrade script
so that it can correctly update the new tax tags on AMLs.

See odoo/odoo#218230
See odoo/enterprise#89912

closes odoo/upgrade#8046

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>