[FIX] im_livechat/saas~18.4: add demo history xmlids
The member history for demo data is now explicitly written
with its own xmlid, and with the new unique constraints on
the member history model, loading the demo xml files will
cause a duplicate entry per channel-partner or channel-guest.
This fix maps the existing demo history to the new xmlids.

https://github.com/odoo/odoo/pull/215228/

closes odoo/upgrade#8057

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>