[FIX] mrp/18.3: set workorder null from stock moves
In V18.3 in Manufacturing order, as per this commit odoo/odoo@13939fb the stock
moves should not be shown if they are not consumed in any operation.However,
during upgrade, moves are stills linked to the workorder in the upgraded
database 18.3, which should not be the case. Hence, while rendering process, it
causes a traceback while accessing workcenters  due to changes in
odoo/enterprise@e80ecb7 as the system attempts to display stock moves associated
 with these orders.

As a result, this commit fixes the issue by setting the workorder_id to null
during the upgrade for stock moves not consumed in any operation, aligning with
the standard behavior in 18.3 and preventing errors during accessing.

Steps to produce
- Create MO
- Open shop floor
- On the created MO, open the menu and open Add Component
- Add any component and close
- Upgrade to V18.3

opw - 4893725
upg - 2999670

closes odoo/upgrade#8060

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>