[FIX] account_reports/saas~18.3.1.0: bypass account_move_line constraint
Some `account.returns` are created with a `closing_move_id` value, which set the
`closing_return_id` of the corresponding `account.move`. The modification of
`closing_return_id` triggers a compute for the `account.move.line.tax_ids`
which trigger the constraint that fails because there is already a `tax_tag_id`
in one `account.move.line`. To bypass the constraint, and because we don't want
`account.move.line.tax_ids` to be recomputed for perf and functional reasons,
this sets the `closing_return_id` in sql.
Depends on this PR:
https://github.com/odoo/upgrade-util/pull/297

closes odoo/upgrade#8079

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>