[IMP] ai_{fields,server_actions}/saas~18.5: migrate to custom rendering
Purpose
=======
Because we need more control on the rendering (to be able to add
the currency for monetary field, to upload attachment in a different
API call, etc). We decided to not use Qweb anymore, because it will
require to many tricks to make it work, and we do the rendering
ourselves, based on LXML, like it has been done for prompt in mail
template.

The new format is
```
<span data-ai-field="partner_id.name">Display Name</span>
<span data-ai-record="27"></span>
```

Task-4915266
See odoo/odoo/pull/218725
See odoo/enterprise/pull/90124

closes odoo/upgrade#8073

Related: odoo/odoo#218725
Related: odoo/enterprise#90124
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>