[IMP] documents_account/saas~18.5: force accounting synchronization
Previously, the user could choose whether to synchronize accounting document or
not. Now it is enabled as soon as the module documents_account is enabled and
the user cannot disable it without uninstalling the module.

In this upgrade, we remove the flags that allowed user to choose to enable
synchronization. Although it is always enabled, it must be configured by journal
to actually work. This is now configured automatically at the creation of the
journal and when installing documents_account. A test
(test_create_journal_without_base_folder) checks that the auto configuration
works (by doing nothing in this case) when the account_folder_id is not set on
the related company. This is why we don't ensure that the finance folder
(documents.document_finance_folder) exists and is set on all companies
(which should be the case) in this upgrade.

See odoo/odoo#219925
See odoo/enterprise#90662

Task-4873102

closes odoo/upgrade#8105

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>