[IMP] l10n_br_avatax/saas~18.5: adapt to the new NCM model constraints
The NCM code model gets a UNIQUE(code, name) constraint. We migrate
potentially duplicated codes by assigning one NCM code per
(code, name) pair and deleting others. While doing this we take care
to prefer keeping codes with an XML ID. Otherwise, the user would end
up with errors when reloading the data.

Because the code and name fields become required we fill them with a
default value (the database ID) to not block the NOT NULL constraint
the ORM will try to set.

We also migrate the name of each NCM code. Before, the code was
prepended to each NCM code name. In master we update the data to not
do that anymore. This migrates existing names to achieve the same
result.

Enterprise PR: https://github.com/odoo/enterprise/pull/91750

task-4998498

closes odoo/upgrade#8178

Related: odoo/enterprise#91750
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>