[FIX] base/saas~18.3: auto-install l10n_be without init
In saas~18.3 the module l10n_be can be autoinstalled as a new
dependency for module l10n_be_account_disallowed_expenses_fleet.
The install on CI with demo will load the belgian CoA and fail
on the post_load for demo data when posting moves before the
company lock dates.
This fix installs the module but skips the demo function calls.

https://github.com/odoo/enterprise/pull/84282

closes odoo/upgrade#8130

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>