[FIX] base/saas~16.4: Fix recycle private partner action
In version 16.4, private addresses have been removed from
the res.partner model. Instead, private address
information has been moved to their respective models.
https://github.com/odoo/upgrade/pull/4678

However, when a customer triggers the server action, archived
records are not transferred to the new table due to issues with
joins. Specifically, if a record in res.partner is missing
the state_id, country_id, or both (Since country_id and state_id
are not required fields, the customer can leave them as null),
it gets excluded from the result set and is therefore not inserted
into the new table.
https://github.com/odoo/upgrade/blob/762c9b48acc971d227d9842167a693dc5aedc235/migrations/base/saas~16.4.1.3/recycle_private_partners.py#L143-L159

here set the field values to null for the archived records marked
with the tag "Old Private Address", which results in data loss,
as the information is not migrated to the new table before being
cleared.
https://github.com/odoo/upgrade/blob/762c9b48acc971d227d9842167a693dc5aedc235/migrations/base/saas~16.4.1.3/recycle_private_partners.py#L171-L188

- OPW: 4971565
- UPG: 3036971

closes odoo/upgrade#8134

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>