[FIX] point_of_sale/17.0: Remove archived pricelists
There's no mechanism to prevent a pricelist from being archived when it's being
used in a PoS config. Archived pricelists are filtered out from the available
pricelists list, so it will trigger a validation error.

The error happens during the chart of accounts load, even for archived configs.

tbg-1724

```
raise ValidationError(_("The default pricelist must be included in the available pricelists."))
```

closes odoo/upgrade#8140

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>