[FIX] product/saas~18.4: fix demo xmlids
The renamed xmlids were still used in other modules.
This fix duplicates the model data to ensure old and
new names exist, similar to the behavior of
'ir.model._update_xmlids'.
It also adds an xmlid for the attribute line of
product8, which is normally deleted when the
line is created from xml file, and the product
xmlid is not recreated because the function being
skipped. The product is used in sale demo data.

closes odoo/upgrade#8155

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>