[IMP] {l10n_be_,}hr_payroll/saas~18.5: move credit time to BE
Related PRs:
- https://github.com/odoo/enterprise/pull/83375
- https://github.com/odoo/odoo/pull/206658

The related PR move the credit time from standard to BE and change the
way we define a version as "credit time".
The upgrade script should get all belgian versions that are credit time.
Then, for each of these versions, create a new calendar that is filling
the previous part time calendar with credit time attendance lines to
match the company calendar. If there is a mismatch, we do nothing.

task-4578909

closes odoo/upgrade#8158

Related: odoo/odoo#206658
Related: odoo/enterprise#83375
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>