[IMP] appointment/saas~18.5: merge single child menu and parent
The reporting menu in the calendar app, (available only when
appointment is installed), has a single child 'All Appointments'.

In order to ease the access to reporting and user experience,
we delete the child menu and link its action to the parent one.

When pressing 'Reporting', it will go the the action previously
linked to the 'All appointments' menu.

COM PR: odoo/odoo#219677
ENT PR: odoo/enterprise#90529

Task-4929768

closes odoo/upgrade#8160

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>