[FIX] account/16.0: Set invoice_date in posted expense journal entries
Before the changes introduced in this pull request (#5771), the journal entries
created by the hr_expense module had their move_type set to 'entry'. However, in
newer versions of Odoo, the expected and default move_type for these kinds of
records is 'in_invoice'.

The purpose of PR #5771 is to update the move_type to 'in_invoice', aligning it
with the behavior in newer versions. However, this change is affecting the
functionality introduced in an earlier pull request (#4341).

Here's the issue:

PR (#5771) modifies the move_type before the logic in PR (#4341) is executed. As
a result, not all records are being handled correctly. Specifically,
the invoice_date is not being set for the records that have their move_type
changed by PR (#5771). This leads to the same problem that PR (#4341) was
intended to fix—missing invoice_date values, which are now required fields.

To resolve this and ensure that no data is skipped or left inconsistent, I’ve
updated the logic to set the invoice_date at the time the move_type is changed.
This ensures that all affected records will have a valid invoice_date.

- create fresh database of v15 and install `hr_expense` module
- create expense => create report => submit to manager => approve =>
     post journal entries
- in the version 15 the invoice_field date is not visible in posted journal
   entries because of its move type and it is not required field.
- now migrate the database to v16
- When you check the journal entries related to this it show empty invoice date
and if you try to update that move you get  user error 'The Bill/Refund date is
required to validate this document.'

closes odoo/upgrade#8173

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>