[IMP] stock_account/saas~18.5: New stock valuation
The inventory valuation has been recreated from start.
The `stock.valuation.layer` object is drop and it's replace by
a field `value` on `stock.move`

The main PR is not totaly finished but it contains all the model
modification.

Since it's needed for the freeze, this PR pass the runbot as green
but it's far from finish.

Still to do:
- Keep a location account in the new field on location (from in or out)
- Compute the new value on the move from the existing layers
- Copy quantity_svl/value_svl to the new existing field
- Create the new variation account and set it as default on stock
  valauation account.

We probably wants to keep the input/output account since we don't want
to impact old valuation.

task-4826464

closes odoo/upgrade#8189

Related: odoo/odoo#222169
Related: odoo/enterprise#91890
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>