[FIX] link_tracker/0.0.0: clean old constraint
A previous fix already tackled the missing removal of a deprecated
constraint link_tracker_url_utms_uniq in 14.4.
https://github.com/odoo/upgrade/commit/06ad72423364e037da8133c4e1fc8e8b1f337ec7

But the issue remains that currently any database that was migrated before
this fix, still carries the constraint.
It will cause issue once a database migrates to 18.0+, where it will break
certain email marketing templates and block sending.

To address the issue, we added a 0.0.0 script that will pre-emptively remove
the constraint for any database migration between version 14.4 and 19.0.
We also remove the redundant script for 14.4

opw-4680129

closes odoo/upgrade#8201

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>