[IMP] appointment/saas~18.5: remove bad configured slots, view and menu
We add a constraint on the appointment.slot model to ensure
start_datetime <= end_datetime as no check was done before.

This only concerns 'unique' slots, as 'recurring' ones do not
use those fields at all.

We can remove them altogether because the slots are only linked
to the appointment type, and used to generate slots in the
booking flow. Removing those will not have any real negative
impact as the verification checks on the slot would fail when
trying to book them in the first place.

Also remove a submenu and a linked view that are not needed and
are replaced with a 'flexible' filter on the main appointment.type
views that the user can use in order to only see its 'custom'
appointments.

ENT PR : odoo/enterprise#91824

Task-4997980

closes odoo/upgrade#8206

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>