[ADD] sign/saas~18.5.1.0: change default_{width, height} storing
Since PR odoo/enterprise#89156 on enterprise gets rid
of the default_{width, height} fields on the form view
of sign.item.type and substitutes them with a
new field field_size, here we set the old fields as non stored.
Also, we set a default value for the new column.

closes odoo/upgrade#7993

Task: 4908814
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>