[IMP] account_transfer,base/saas~18.5: Automatic transfers revamp
Description of the issue this commit addresses:

The automatic transfers have been the way they are for a long time now and are
barely used compared to how they could be. A little revamp could give them a
bigger impact and use.

Desired behavior after this commit is merged:

The automatic transfers have a renewed UI and use different systems to redirect
the amounts of the original moves.

enteprise pr: https://github.com/odoo/enterprise/pull/87015
task-4813376

closes odoo/upgrade#7887

Related: odoo/enterprise#87015
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>