[FIX] stock_delivery/saas~16.2: move fields
In odoo/odoo#110686 the fields for `stock.move.line` in `delivery`
module where moved to `stock_delivery`. This was not reflected in the
upgrade PR.

When upgrading from 16.0 it may be that we force install delivery in
16.2. The script in delivery/16.1 won't run. We need to create
`carrier_id` column for stock move lines. In principle only pickings
created by demo data get a value already set for carrier_id. Still, to
avoid any issue we still compute the column in post- of stock_delivery.

We also improve the original query to update only if it writes not null
values.