[FIX] l10n_in_reports/saas~18.5: update return_period_month_year
In this commit:
https://github.com/odoo/enterprise/commit/7694e8eb5c041528fa5c7b9036408fc69669483a
We managed the write computation of the `return_period_month_year` field.
Compute change will work for the new records, not for the existing ones directly
and for that, we added the temp code to recompute the field.

Temporary code:
https://github.com/odoo/enterprise/blob/4f867f4549d2a21d2819b4169677ba1612330f7a/l10n_in_reports/models/gst_return_period.py#L1229-L1234
for saas-18.3 and saas-18.4

And in the master, we removed that code and kept only the compute method change.
fw of the PR where temporary code was removed
https://github.com/odoo/enterprise/commit/b2ff9de5bd535174688edbfc1553d9f3fcb5d02d
But migration remained for the removed code, which will be done with this
commit.

closes odoo/upgrade#8236

Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>