[FIX] l10n_es_edi_verifactu_pos/saas~18.5: remove cancel related field
In the community PR we remove the possibility to cancel PoS orders.
I.e. we remove the button from the form view. The visibility of the button
was controlled by computed field `l10n_es_edi_verifactu_show_cancel_button`.
The removal of the button made the field obsolete. It is removed in this PR.

See odoo/odoo#223050

closes odoo/upgrade#8243

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>